experimenting with network based flow control
This may well be considered a failure as it is not reliable
enough. The whole Port stuff in Erlang is definetely not made
for large input and there does not seem to be any project that
would allow {active, once} semantics on the output of a port
command.

My best bet is to rework glacier-server into a glacier-proxy
that would act as a RESTful HTTP frontend to Glacier but with
a much simpler API that can be used with curl.